+++ src/declarations.d.ts
@@ -1,0 +1,10 @@
+export {};
+
+declare global {
+  interface Window {
+    plausible: (
+      name: string,
+      options: { callback?: () => void; props?: { [key: string]: string } }
+    ) => void;
+  }
+}